[tdm] refactor: make trajectory runtime reusable
Centralize deterministic boundary guards and let TDM-R1 inherit TDM's shared math so algorithm files retain only their distinct objectives and reward behavior.
